build: ihe gw configs on the docker image
Ref. metriport/metriport-internal#1350

+++ packages/ihe-gateway/scripts/pull-from-server.sh
@@ -4,7 +4,11 @@
 set -e
 
 source ./scripts/load-env.sh
-source ./scripts/check-env.sh
+
+if [ -z "${IHE_GW_FULL_BACKUP_LOCATION}" ]; then
+  echo "Error: IHE_GW_FULL_BACKUP_LOCATION is not set, exiting"
+  exit 1
+fi
 
 # clear out the server directory because mirthsync doesn't remove files that are no longer on the server
 rm -rf ./server/Channels
@@ -16,3 +20,5 @@ rm -rf ./server/GlobalScripts
 
 # "-m" flag: only the FullBackup.xml file, "Equivalent to Mirth Administrator backup and restore"
 ./scripts/mirthsync.sh -s $IHE_GW_URL -u $IHE_GW_USER -p $IHE_GW_PASSWORD -i -t $IHE_GW_FULL_BACKUP_LOCATION -m backup -f pull
+
+mv $IHE_GW_FULL_BACKUP_LOCATION/FullBackup.xml $IHE_GW_FULL_BACKUP_LOCATION/FullBackup-$ENV_TYPE.xml
